Update 2021-03-18-li21a.md
Hi,

I'm Yuantong Li, one of the authors of the paper "Online Forgetting Process for Linear Regression Models". We want to make the following changes:

There is a mistake in the appendix' lemmas and theorems' indexes. We changed the paper and its appendix paper.

There are mistakes in the of both camera ready and supplement, we attached the updated files with correct index.

Thank you so much for your help!